requirements.txt
requests=2.22.0
mysql-connector-python==8.0.18

